fix: enable_mock
