+++ src/pages/content/ui/app.test.tsx
@@ -1,6 +1,6 @@
 import { describe, test } from 'vitest';
 import { render, screen } from '@testing-library/react';
-import App from '@root/src/pages/content/ui/App';
+import App from '@root/src/pages/content/ui/app';
 
 describe('appTest', () => {
   test('render text', () => {
