+++ src/pages/content/ui/root.tsx
@@ -1,5 +1,5 @@
 import { createRoot } from 'react-dom/client';
-import App from '@root/src/pages/content/ui/App';
+import App from '@root/src/pages/content/ui/app';
 import refreshOnUpdate from 'virtual:reload-on-update-in-view';
 import injectedStyle from './injected.css?inline';
 refreshOnUpdate('pages/content');
